Introduce initial version of register abstraction layer
This commit introduces an abstraction layer over writing to
registers. See the update hello/hi_rust.rs file for an example.

The compute() and write() methods are marked as #[inline(always)]
because only then, it appears, can LLVM optimize the calls away
properly, because the values in are constants known at compile time,
the calls to write() should not generate any more assembly than a
direct call to volatile_store(). Without these annotations, the
binary grows to almost 1600 bytes and starts requiring memcpy.

I really don't know why this is the case...